+++ src/main/java/chess/domain/Color.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain;
 
 public enum Color {
     WHITE,
