+++ src/test/java/chess/domain/PieceTest.java
@@ -1,6 +1,5 @@
 package chess.domain;
 
-import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.position.Position;
