+++ src/test/java/chess/domain/piece/RookTest.java
@@ -3,6 +3,7 @@ package chess.domain.piece;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import chess.domain.Color;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
